Added test for data. Modified csv.
